Create prototype API
