+++ environment.yaml
@@ -1,0 +1,162 @@
+name: speechgui
+channels:
+  - pytorch
+  - nvidia
+  - conda-forge
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=conda_forge
+  - _openmp_mutex=4.5=2_kmp_llvm
+  - arrow=1.2.3=py311h06a4308_1
+  - asgiref=3.5.2=py311h06a4308_0
+  - asttokens=2.0.5=pyhd3eb1b0_0
+  - backcall=0.2.0=pyhd3eb1b0_0
+  - blas=1.0=mkl
+  - brotli-python=1.0.9=py311h6a678d5_8
+  - bzip2=1.0.8=h5eee18b_6
+  - ca-certificates=2024.3.11=h06a4308_0
+  - cached-property=1.5.2=py_0
+  - certifi=2024.2.2=py311h06a4308_0
+  - charset-normalizer=2.0.4=pyhd3eb1b0_0
+  - colorama=0.4.6=py311h06a4308_0
+  - cuda-cudart=12.1.105=0
+  - cuda-cupti=12.1.105=0
+  - cuda-libraries=12.1.0=0
+  - cuda-nvrtc=12.1.105=0
+  - cuda-nvtx=12.1.105=0
+  - cuda-opencl=12.4.127=0
+  - cuda-runtime=12.1.0=0
+  - django=4.1=py311h06a4308_0
+  - dtw=1.4.0=pyhd8ed1ab_0
+  - executing=0.8.3=pyhd3eb1b0_0
+  - ffmpeg=4.2.2=h20bf706_0
+  - filelock=3.13.1=py311h06a4308_0
+  - fqdn=1.5.1=pyhd8ed1ab_0
+  - freetype=2.12.1=h4a9f257_0
+  - gettext=0.22.5=h59595ed_2
+  - gettext-tools=0.22.5=h59595ed_2
+  - gmp=6.2.1=h295c915_3
+  - gmpy2=2.1.2=py311hc9b5ff0_0
+  - gnutls=3.6.15=he1e5248_0
+  - idna=3.7=py311h06a4308_0
+  - intel-openmp=2023.1.0=hdb19cb5_46306
+  - isoduration=20.11.0=pyhd8ed1ab_0
+  - jedi=0.18.1=py311h06a4308_1
+  - jinja2=3.1.3=py311h06a4308_0
+  - jpeg=9e=h5eee18b_1
+  - jsonpointer=2.1=pyhd3eb1b0_0
+  - lame=3.100=h7b6447c_0
+  - lcms2=2.12=h3be6417_0
+  - ld_impl_linux-64=2.38=h1181459_1
+  - lerc=3.0=h295c915_0
+  - libasprintf=0.22.5=h661eb56_2
+  - libasprintf-devel=0.22.5=h661eb56_2
+  - libcublas=12.1.0.26=0
+  - libcufft=11.0.2.4=0
+  - libcufile=1.9.1.3=0
+  - libcurand=10.3.5.147=0
+  - libcusolver=11.4.4.55=0
+  - libcusparse=12.0.2.55=0
+  - libdeflate=1.17=h5eee18b_1
+  - libffi=3.4.4=h6a678d5_1
+  - libflac=1.4.3=h59595ed_0
+  - libgcc-ng=13.2.0=h77fa898_7
+  - libgettextpo=0.22.5=h59595ed_2
+  - libgettextpo-devel=0.22.5=h59595ed_2
+  - libgfortran-ng=11.2.0=h00389a5_1
+  - libgfortran5=11.2.0=h1234567_1
+  - libidn2=2.3.4=h5eee18b_0
+  - libjpeg-turbo=2.0.0=h9bf148f_0
+  - libnpp=12.0.2.50=0
+  - libnvjitlink=12.1.105=0
+  - libnvjpeg=12.1.1.14=0
+  - libogg=1.3.5=h27cfd23_1
+  - libopus=1.3.1=h7b6447c_0
+  - libpng=1.6.43=h2797004_0
+  - libsndfile=1.2.2=hc60ed4a_1
+  - libstdcxx-ng=13.2.0=hc0a3c3a_7
+  - libtasn1=4.19.0=h5eee18b_0
+  - libtiff=4.5.1=h6a678d5_0
+  - libunistring=0.9.10=h27cfd23_0
+  - libuuid=1.41.5=h5eee18b_0
+  - libvorbis=1.3.7=h7b6447c_0
+  - libvpx=1.7.0=h439df22_0
+  - libwebp-base=1.3.2=h5eee18b_0
+  - libzlib=1.2.13=hd590300_5
+  - llvm-openmp=14.0.6=h9e868ea_0
+  - lz4-c=1.9.4=h6a678d5_1
+  - mad=0.15.1b=h9c3ff4c_1
+  - markupsafe=2.1.3=py311h5eee18b_0
+  - mkl=2023.1.0=h213fc3f_46344
+  - mkl-service=2.4.0=py311h5eee18b_1
+  - mkl_fft=1.3.8=py311h5eee18b_0
+  - mkl_random=1.2.4=py311hdb19cb5_0
+  - mpc=1.1.0=h10f8cd9_1
+  - mpfr=4.0.2=hb69a4c5_1
+  - mpg123=1.32.6=h59595ed_0
+  - mpmath=1.3.0=py311h06a4308_0
+  - ncurses=6.4=h6a678d5_0
+  - nettle=3.7.3=hbbd107a_1
+  - networkx=3.1=py311h06a4308_0
+  - numpy=1.26.4=py311h08b1b3b_0
+  - numpy-base=1.26.4=py311hf175353_0
+  - openh264=2.1.1=h4ff587b_0
+  - openjpeg=2.4.0=h3ad879b_0
+  - openssl=3.0.13=h7f8727e_1
+  - parso=0.8.3=pyhd3eb1b0_0
+  - pexpect=4.8.0=pyhd3eb1b0_3
+  - pickleshare=0.7.5=pyhd3eb1b0_1003
+  - pillow=10.3.0=py311h5eee18b_0
+  - pip=24.0=py311h06a4308_0
+  - portaudio=19.6.0=h7b6447c_4
+  - prompt-toolkit=3.0.43=py311h06a4308_0
+  - ptyprocess=0.7.0=pyhd3eb1b0_2
+  - pure_eval=0.2.2=pyhd3eb1b0_0
+  - pyaudio=0.2.11=py311h5eee18b_2
+  - pysocks=1.7.1=py311h06a4308_0
+  - python=3.11.9=h955ad1f_0
+  - python-dateutil=2.9.0post0=py311h06a4308_0
+  - pytorch=2.3.0=py3.11_cuda12.1_cudnn8.9.2_0
+  - pytorch-cuda=12.1=ha16c6d3_5
+  - pytorch-mutex=1.0=cuda
+  - pyyaml=6.0.1=py311h5eee18b_0
+  - readline=8.2=h5eee18b_0
+  - requests=2.31.0=py311h06a4308_1
+  - scipy=1.12.0=py311h08b1b3b_0
+  - setuptools=69.5.1=py311h06a4308_0
+  - shellingham=1.5.0=py311h06a4308_0
+  - six=1.16.0=pyhd3eb1b0_1
+  - sox=14.4.2=h32e7c5b_1019
+  - sqlite=3.45.3=h5eee18b_0
+  - sqlparse=0.4.4=py311h06a4308_0
+  - stack_data=0.6.2=pyhd8ed1ab_0
+  - sympy=1.12=py311h06a4308_0
+  - tbb=2021.8.0=hdb19cb5_0
+  - tk=8.6.14=h39e8969_0
+  - torchaudio=2.3.0=py311_cu121
+  - torchtriton=2.3.0=py311
+  - torchvision=0.18.0=py311_cu121
+  - typing_extensions=4.11.0=py311h06a4308_0
+  - tzdata=2024a=h04d1e81_0
+  - uri-template=1.3.0=pyhd8ed1ab_0
+  - urllib3=2.2.1=py311h06a4308_0
+  - wcwidth=0.2.5=pyhd3eb1b0_0
+  - webcolors=1.13=pyhd8ed1ab_0
+  - wheel=0.43.0=py311h06a4308_0
+  - x264=1!157.20191217=h7b6447c_0
+  - xz=5.4.6=h5eee18b_1
+  - yaml=0.2.5=h7b6447c_0
+  - yarg=0.1.9=py_1
+  - zlib=1.2.13=hd590300_5
+  - zstd=1.5.5=hc292b87_2
+  - pip:
+      - conformer==0.3.2
+      - einops==0.8.0
+      - fsspec==2024.5.0
+      - llvmlite==0.42.0
+      - more-itertools==10.2.0
+      - numba==0.59.1
+      - openai-whisper==20231117
+      - regex==2024.5.15
+      - tiktoken==0.7.0
+      - tqdm==4.66.4
